PR reviews
Commenting classes
Fixing README
