+++ 1-Integrated-Cache/1-1-WebApp-BgWorker/BackgroundWorker/MsalAccount.cs
@@ -2,7 +2,9 @@
 
 namespace BackgroundWorker
 {
-    // TODO: Comment, very well
+    // Msal expects an object of IAccount in the method AcquireTokenSilent. 
+    // This class implements the IAccount interface and will be instantiated with the values that came from MsalAccountActivity entity, and
+    // will be sent as parameter to AcquireTokenSilent, so Msal can grab to cached token for that particular user account.
     public class MsalAccount : IAccount
     {
         public MsalAccount() { }
