+++ 1-Integrated-Cache/1-1-WebApp-BgWorker/WebApp/Services/GraphServiceClientFactory.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 
 namespace WebApp.Services
 {
-    // TODO: Comment
+    // Factory to instantiate the GraphServiceClient to be used to call Graph 
     public class GraphServiceClientFactory
     {
         public static GraphServiceClient GetAuthenticatedGraphClient(Func<Task<string>> acquireAccessToken,
