+++ 1-Integrated-Cache/1-2-WebAPI-BgWorker/BackgroundWorker/MsalAccount.cs
@@ -2,6 +2,9 @@
 
 namespace BackgroundWorker
 {
+    // Msal expects an object of IAccount in the method AcquireTokenSilent. 
+    // This class implements the IAccount interface and will be instantiated with the values that came from MsalAccountActivity entity, and
+    // will be sent as parameter to AcquireTokenSilent, so Msal can grab to cached token for that particular user account.
     public class MsalAccount : IAccount
     {
         public MsalAccount() { }
