--- 1-Integrated-Cache/1-2-WebAPI-BgWorker/WebAPI/Models/TodoItem.cs
@@ -1,11 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-namespace WebAPI.Models
-{
-    public class TodoItem
-    {
-        public string Owner { get; set; }
-        public string Title { get; set; }
-    }
-}
